tests/requirements.txt
pytest
pytest-asyncio
pytest-mock
requests
httpx
sqlalchemy
alembic
python-dotenv